Add more patterns to myscelium wrapper and also implement the basic redirect mecanism in the tranposer & refactory some code to make it more cleaner and safe to a range of possible errors conditions and update some methos to the correct ones!

+++ Myscelium/src/socket_host/enhanced_buffer/buffer_down_mananger.rs
@@ -72,7 +72,7 @@ pub fn set_workers_num (n_workers:u32) {
     pub command_id:i32,
     pub client_id:String,
     pub parity_id:String,
-    pub priority:i32,
+    pub priority:u8,
     pub command:String,
 }
